# swagger.yaml
swagger: "2.0"
info:
  description: "Server Definition for the Governor. "
  version: "1.0.0"
  title: "Governor"
  termsOfService: "http://servec.com/terms/"
  contact:
    email: "ivamsiravi@gmail.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "servec.com"
basePath: "/v2"
...
schemes:
- "https"
- "http"
paths:
  /triggers:
    post:
      tags: 
      - "trigger"
      summary: "Create New Trigger"
      operationId: "addTrigger"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Trigger object"
        required: true
        schema:
          $ref: "#/definitions/Trigger"
      responses:
        201:
          description: "Created"
          schema:
            $ref: "#/definitions/Trigger"
        400: 
          description: "Bad Request"
          schema: 
            $ref: "#/definitions/BadRequest"
        409:
          description: "Conflict"
    get:
      tags:
      - "trigger"
      summary: "Retrieve All Triggers"
      operationId: "getTriggers"
      responses:
        200:
          description: OK
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Trigger"
  /triggers/{triggerId}:
    put:
      tags:
      - "role"
      summary: "Update existing Trigger"
      operationId: "updateTriggerById"
      parameters:
      - name: "triggerId"
        in: "path"
        required: true
        type: "integer"
        format: "int64"
      responses:
        204:
          description: "No Content"
          schema:
            $ref: "#/definitions/Trigger"
        400: 
          description: "Bad Request"
          schema: 
            $ref: "#/definitions/BadRequest"
        404: 
          description: "Not Found"
    delete:
      tags:
      - "trigger"
      summary: "Delete existing Trigger"
      operationId: "deleteTriggerById"
      parameters:
      - name: "triggerId"
        in: "path"
        description: ""
        required: true
        type: "integer"
        format: "int64"
      responses:
        204:
          description: "No Content"
          schema:
            $ref: "#/definitions/Trigger"
        404: 
          description: "Not Found"
    get:
      tags:
      - "trigger"
      summary: "Retrieve Existing Trigger"
      operationId: "getTriggerById"
      parameters:
      - name: "triggerId"
        in: "path"
        description: ""
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/Trigger"
        404: 
          description: "Not Found"
...
  /users:
    post:
      tags:
      - "user"
      summary: "Create user"
      description: "This can only be done by the logged in user."
      operationId: "createUser"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Created user object"
        required: true
        schema:
          $ref: "#/definitions/User"
      responses:
        default:
          description: "successful operation"
  /user/login:
    post:
      tags:
      - "user"
      summary: "Logs user into the system"
      description: ""
      operationId: "loginUser"
      produces:
      - "application/json"
      parameters:
      - name: "body"
        in: "body"
        required: true
        schema:
          $ref: "#/definitions/Login"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/User"
        400:
          description: "Invalid username/password supplied"
  /user/{username}/logout:
    put:
      tags:
      - "user"
      summary: "Logs out current logged in user session"
      description: ""
      operationId: "logoutUser"
      produces:
      - "application/json"
      parameters: 
      - name: "username"
        in: "path"
        description: "The user name for login"
        required: true
        type: "string"
      responses:
        204:
          description: "Accepted"
          schema:
            $ref: "#/definitions/User"
        400: 
          description: "Bad Request"
          schema: 
            $ref: "#/definitions/BadRequest"
        404: 
          description: "Not Found"
  /users/{username}:
    get:
      tags:
      - "user"
      summary: "Get user by user name"
      description: ""
      operationId: "getUserByName"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "username"
        in: "path"
        description: "The name that needs to be fetched. Use user1 for testing. "
        required: true
        type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/User"
        400:
          description: "Invalid username supplied"
        404:
          description: "User not found"
    put:
      tags:
      - "user"
      summary: "Updated user"
      description: "This can only be done by the logged in user."
      operationId: "updateUser"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "username"
        in: "path"
        description: "name that need to be updated"
        required: true
        type: "string"
      - in: "body"
        name: "body"
        description: "Updated user object"
        required: true
        schema:
          $ref: "#/definitions/User"
      responses:
        400:
          description: "Invalid user supplied"
        404:
          description: "User not found"
    delete:
      tags:
      - "user"
      summary: "Delete user"
      description: "This can only be done by the logged in user."
      operationId: "deleteUser"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "username"
        in: "path"
        description: "The name that needs to be deleted"
        required: true
        type: "string"
      responses:
        400:
          description: "Invalid username supplied"
        404:
          description: "User not found"
...
definitions:
...
  Trigger:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      name:
        type: "string"
      enabled:
        type: "boolean"
...
  BadRequest:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      code:
        type: "integer"
        format: "int32"
      message:
        type: "string"
      fields:
        type: "array"
        items:
          type: "string"
  Login:
    type: "object"
    properties:
      username: 
        type: "string"
      password:
        type: "string"
  User:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      username:
        type: "string"
      firstName:
        type: "string"
      lastName:
        type: "string"
      email:
        type: "string"
      password:
        type: "string"
      phone:
        type: "string"
      tier:
        type: "string"
        enum:
        - "enterprise"
        - "free"
      status:
        type: "string"
        enum:
        - "active"
        - "pending"
        - "suspended"